+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ CHANGELOG
 
 * Added: Support for reading vulns and asset exports into memory
 * Added: Support for graceful back-off
+* Added: Support for tagging API
 
 1.4.1
 ==========